created new logger file for test cases so change can be done incrementally
